+++ docs/history.rst
@@ -2,11 +2,14 @@
 History and release notes
 =========================
 
+.. currentmodule:: parsy
+
 0.9.0 - unreleased
 ------------------
 
 * Better error reporting of failed parses.
 * Documentation overhaul and expansion.
+* Added :meth:`Parser.combine`.
 
 0.0.4 - 2014-12-28
 ------------------